feat: get orders & users for admin

+++ Models/Product.cs
@@ -5,6 +5,6 @@ namespace SportShop.Models
         public string Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }
-    }   
+        public double Price { get; set; }
+    }
 }
